// src/types.d.ts
export interface Context {
  listDepth: number
  listOrdered: boolean
  listItemIndex: number
  currentContext?: string
  svg: boolean
}

export type Attributes = Record<string, string>

interface TTableProps {
  /** whether it is in head */
  isHead: boolean
}

interface THeadingProps {
  /** heading level */
  level: number
}

interface TListProps {
  /** depth of the list */
  depth: number
  /** whether it is ordered list */
  ordered: boolean
}

interface TCodeProps {
  /** language name original @example 'language-js' */
  languageOriginal: string

  /** language name @example 'js' */
  language: string

  /** code content */
  content: string

  /** whether it is inline code */
  inline: boolean
}

// https://www.google.com/search?q=record%3Cstring,+any%3E+vs+record%3Cstring,+unknown%3E&sourceid=chrome&ie=UTF-8
export type CustomAttrsObjectResult = Record<string, unknown>

type CustomAttrsFunctionValue<T> = (
  /**
   * hast node
   *
   * Please refer to the source code at the following URL to understand the possible attributes for each element.
   *
   * @see https://github.com/syntax-tree/mdast-util-to-hast/tree/main/lib/handlers
   */
  node: Element,
  /**
   * Properties of the current element.
   *
   * Except for the basic properties provided from hast, it also includes custom properties such as `level`, `ordered`, `depth`, `index` etc.
   */
  combinedAttrs: T | Attributes
) => Record<string, unknown>

type CustomAttrsValue<T extends Record<string, unknown> = Record<string, unknown>> =
  | CustomAttrsObjectResult
  | CustomAttrsFunctionValue<T>

type TBasicHTMLTagNames = keyof Omit<
  HTMLElementTagNameMap,
  'h1' | 'h2' | 'h3' | 'h4' | 'h5' | 'h6' | 'ul' | 'ol' | 'li' | 'code' | 'td' | 'th' | 'tr'
>
export type CustomAttrs = {
  [key in TBasicHTMLTagNames]?: CustomAttrsValue // << dynamic properties
} & {
  [key: string]:
    | CustomAttrsValue
    | CustomAttrsValue<THeadingProps>
    | CustomAttrsValue<TListProps>
    | CustomAttrsValue<TCodeProps>
    | CustomAttrsValue<TTableProps>
    | undefined
  ['h1']?: CustomAttrsValue<THeadingProps> // << static properties
  ['h2']?: CustomAttrsValue<THeadingProps>
  ['h3']?: CustomAttrsValue<THeadingProps>
  ['h4']?: CustomAttrsValue<THeadingProps>
  ['h5']?: CustomAttrsValue<THeadingProps>
  ['h6']?: CustomAttrsValue<THeadingProps>
  ['heading']?: CustomAttrsValue<THeadingProps>
  ['ul']?: CustomAttrsValue<TListProps>
  ['ol']?: CustomAttrsValue<TListProps>
  ['list']?: CustomAttrsValue<TListProps>
  ['li']?: CustomAttrsValue<TListProps>
  ['list-item']?: CustomAttrsValue<TListProps>
  ['code']?: CustomAttrsValue<TCodeProps>
  ['inline-code']?: CustomAttrsValue<TCodeProps>
  ['block-code']?: CustomAttrsValue<TCodeProps>
  ['td']?: CustomAttrsValue<TTableProps>
  ['th']?: CustomAttrsValue<TTableProps>
  ['tr']?: CustomAttrsValue<TTableProps>
}

export type AliasList = string[]